.history/requirements_20250814225123.txt
# Core Framework
fastapi
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Telegram
telethon==1.32.0
cryptg==0.4.0

# Image Processing
Pillow==10.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Utils
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
faker==20.1.0

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.0
mypy==1.7.1
pre-commit==3.5.0
